Initial check in of AS3 Utilities
I am adding KDateTime class. This is a great improvement over the
standard AS3 class 'Date'. It is a reverse engineered version of the
.NET DateTime class and is compatible with that class ( in the sense
that the .Ticks member is identical in both ).
The Date class in flash and the CRT Date/Time are both extrememly
limited in that they measure time in MS from 1/1/1970. This class
measures time in 100 nanosecond intervals from 0/0/0000 to 12/31/9999.
It is especially useful in conjunction with an ASP backend since you can
exchange date/time using a single long value ( i.e. the .Tick member
common to both classes )